+++ tests/testthat/test-git2r.R
@@ -46,12 +46,12 @@ checkout(r, "master")
 
 test_that("head returns a branch", {
   h <- git2r::repository_head(r)
-  expect_identical(class(h), "git_branch")
+  expect_true(inherits(h, "git_branch"))
 })
 
 test_that("lookup returns a commit", {
   c_look <- lookup(r, c1$sha)
-  expect_identical(class(c_look), "git_commit")
+  expect_true(inherits(c_look, "git_commit"))
   expect_identical(c_look, c1)
 })
 
@@ -62,34 +62,34 @@ test_that("workdir does not includes a trailing slash", {
 
 test_that("as.list returns a list", {
   l <- as.list(t1)
-  expect_identical(class(l), "list")
+  expect_true(inherits(l, "list"))
 })
 
 test_that("as.data.frame returns a data.frame", {
   d <- as.data.frame(t1)
-  expect_identical(class(d), "data.frame")
+  expect_true(inherits(d, "data.frame"))
 })
 
 test_that("diff returns a diff", {
   difference <- base::diff(t1, t2)
-  expect_identical(class(difference), "git_diff")
+  expect_true(inherits(difference, "git_diff"))
   diff_file_list <- difference$files
-  expect_identical(class(diff_file_list), "list")
+  expect_true(inherits(diff_file_list, "list"))
   diff_file <- diff_file_list[[1]]
-  expect_identical(class(diff_file), "git_diff_file")
+  expect_true(inherits(diff_file, "git_diff_file"))
   new_file <- diff_file$new_file
-  expect_identical(class(new_file), "character")
+  expect_true(inherits(new_file, "character"))
 })
 
 test_that("can access object elements using $", {
   expect_identical(c1$message, "commit 1")
-  expect_identical(class(c1$sha), "character")
+  expect_true(inherits(c1$sha, "character"))
   expect_identical(b2$name, "b2")
 })
 
 test_that("merge returns git_merge_result", {
   m1 <- git2r_merge(r, "b2")
-  expect_identical(class(m1), "git_merge_result")
+  expect_true(inherits(m1, "git_merge_result"))
   expect_identical(m1$up_to_date, FALSE)
   expect_identical(m1$fast_forward, TRUE)
   expect_identical(m1$conflicts, FALSE)
@@ -97,7 +97,7 @@ test_that("merge returns git_merge_result", {
 
   # Merge a second time, which has no effect
   m2 <- git2r_merge(r, "b2")
-  expect_identical(class(m2), "git_merge_result")
+  expect_true(inherits(m2, "git_merge_result"))
   expect_identical(m2$up_to_date, TRUE)
   expect_identical(m2$fast_forward, FALSE)
   expect_identical(m2$conflicts, FALSE)
@@ -105,7 +105,7 @@ test_that("merge returns git_merge_result", {
 
   # Merge parallel branch, which should create a merge commit
   m3 <- git2r_merge(r, "b3")
-  expect_identical(class(m3), "git_merge_result")
+  expect_true(inherits(m3, "git_merge_result"))
   expect_identical(m3$up_to_date, FALSE)
   expect_identical(m3$fast_forward, FALSE)
   expect_identical(m3$conflicts, FALSE)
@@ -115,7 +115,7 @@ test_that("merge returns git_merge_result", {
   # Merge conflicting branch, which should create a merge conflict
   # fail = TRUE
   m4 <- git2r_merge(r, "b4", fail = TRUE)
-  expect_identical(class(m4), "git_merge_result")
+  expect_true(inherits(m4, "git_merge_result"))
   expect_identical(m4$up_to_date, FALSE)
   expect_identical(m4$fast_forward, FALSE)
   expect_identical(m4$conflicts, TRUE)
@@ -124,7 +124,7 @@ test_that("merge returns git_merge_result", {
   expect_identical(status(r)$unstaged, structure(list(), .Names = character(0)))
   # fail = FALSE
   m5 <- git2r_merge(r, "b4", fail = FALSE)
-  expect_identical(class(m5), "git_merge_result")
+  expect_true(inherits(m5, "git_merge_result"))
   expect_identical(m5$up_to_date, FALSE)
   expect_identical(m5$fast_forward, FALSE)
   expect_identical(m5$conflicts, TRUE)